# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BlastRadiusResponse": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "enum": [
              "forward",
              "reverse"
            ],
            "title": "Direction",
            "type": "string"
          },
          "entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "max_hops": {
            "minimum": 1.0,
            "title": "Max Hops",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "dependency",
              "traffic"
            ],
            "title": "Mode",
            "type": "string"
          },
          "relation_type": {
            "$ref": "#/components/schemas/TopologyRelation"
          },
          "root_entity_id": {
            "title": "Root Entity Id",
            "type": "string"
          }
        },
        "required": [
          "root_entity_id",
          "mode",
          "relation_type",
          "direction",
          "max_hops",
          "entity_ids"
        ],
        "title": "BlastRadiusResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "TopologyEdge": {
        "additionalProperties": false,
        "properties": {
          "relation_type": {
            "$ref": "#/components/schemas/TopologyRelation"
          },
          "relationship": {
            "title": "Relationship",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "enum": [
                  "impact_path",
                  "blast_radius"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "relation_type",
          "relationship"
        ],
        "title": "TopologyEdge",
        "type": "object"
      },
      "TopologyNode": {
        "additionalProperties": false,
        "properties": {
          "criticality": {
            "title": "Criticality",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "enum": [
                  "suspected_root",
                  "primary_affected",
                  "impact_path",
                  "blast_radius"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "service",
          "criticality"
        ],
        "title": "TopologyNode",
        "type": "object"
      },
      "TopologyPathResponse": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "enum": [
              "forward",
              "reverse"
            ],
            "title": "Direction",
            "type": "string"
          },
          "distance": {
            "minimum": 0.0,
            "title": "Distance",
            "type": "integer"
          },
          "entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "relation_type": {
            "$ref": "#/components/schemas/TopologyRelation"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "relation_type",
          "direction",
          "distance",
          "entity_ids"
        ],
        "title": "TopologyPathResponse",
        "type": "object"
      },
      "TopologyRelation": {
        "enum": [
          "depends_on",
          "sends_traffic_to"
        ],
        "title": "TopologyRelation",
        "type": "string"
      },
      "TopologySnapshot": {
        "additionalProperties": false,
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/TopologyEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "fixture_version": {
            "title": "Fixture Version",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "fixture_version",
          "nodes",
          "edges"
        ],
        "title": "TopologySnapshot",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Milestone-0 contract API; feature routes are owner-labelled stubs.",
    "title": "Network Anomaly Root-Cause Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/topology": {
      "get": {
        "operationId": "topology_api_v1_topology_get",
        "parameters": [
          {
            "in": "query",
            "name": "incident_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologySnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Topology",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/topology/blast-radius/{entity_id}": {
      "get": {
        "operationId": "blast_radius_api_v1_topology_blast_radius__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": true,
            "schema": {
              "enum": [
                "dependency",
                "traffic"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_hops",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 20,
              "minimum": 1,
              "title": "Max Hops",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlastRadiusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Blast Radius",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/topology/path": {
      "get": {
        "operationId": "path_api_v1_topology_path_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Target",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "relation_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TopologyRelation"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": true,
            "schema": {
              "enum": [
                "forward",
                "reverse"
              ],
              "title": "Direction",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyPathResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Path",
        "tags": [
          "topology"
        ]
      }
    }
  }
}